toolkit/mozapps/update/tests/chrome/test_0073_notify_verifyFailPartialComplete.xul

<?xml version="1.0"?>
<!--
/* Any copyright is dedicated to the Public Domain.
 * http://creativecommons.org/publicdomain/zero/1.0/
 */
-->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Update Wizard pages: errors (partial and complete patches with invalid sizes)"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="runTestDefault();">
<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
<script type="application/javascript"
        src="utils.js"/>

<script type="application/javascript">
<![CDATA[

const TESTS = [ {
  pageid: PAGEID_ERRORS,
  buttonClick: "finish"
} ];

function runTest() {
  debugDump("entering");

  let patches = getLocalPatchString("partial", null, null, null, "1234", null,
                                    STATE_DOWNLOADING) +
                getLocalPatchString("complete", null, null, null, "1234",
                                    "false");
  let updates = getLocalUpdateString(patches, null, null, null,
                                     Services.appinfo.version, null,
                                     null, null, null, null, "false");

  writeUpdatesToXMLFile(getLocalUpdatesXMLString(updates), true);

  writeUpdatesToXMLFile(getLocalUpdatesXMLString(""), false);
  writeStatusFile(STATE_DOWNLOADING);
  reloadUpdateManagerData();

  testPostUpdateProcessing();
}

]]>
</script>

<body xmlns="http://www.w3.org/1999/xhtml">
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>
</body>
</window>
